feat: use monitorEventLoopDelay where available
Use the `monitorEventLoopDelay` API available from Node 12 and up to
sample the event loop delay.
When available, the resolution is hard-coded at 10 ms and all the values of the
`sampleInterval` option lower than that are ignored and its value is set to
10. Also, when using `monitorEventLoopDelay` the default sampleInterval is set
to 1000 since it used only to gather memory metrics.
Tests have been modified to:
1. do not rely on the event loop delay where is not necessary to trigger
   503 because it has been proven unreliable (maybe because I am using the mean
   value from the `Histogram` returned by `monitorEventLoopDelay`)
2. add a `wait` function in tests to wait for event loop samples get
   taken and to wait for the increased default `sampleInterval`

Fix #18.